+++ internal/config/config.go
@@ -8,7 +8,7 @@ type Config struct {
 	GitlabToken    string `envconfig:"GITLAB_TOKEN" required:"true"`
 	GitlabURL      string `envconfig:"GITLAB_URL" default:"https://gitlab.com"`
 	ProjectID      string `envconfig:"GITLAB_PROJECT_ID" required:"true"`
-	MergeRequestID int    `envconfig:"GITLAB_MR_ID" required:"true"`
+	MergeRequestID int64  `envconfig:"GITLAB_MR_ID" required:"true"`
 }
 
 func Load() (*Config, error) {
